+++ packages/client/src/components/forms/LoginForm/LoginForm.scss
@@ -5,8 +5,10 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  row-gap: 1rem;
+  row-gap: 2rem;
 }
 .form-item {
   width: 100%;
+  height: 100%;
+  margin-bottom: 0;
 }
